created a separate occupancy repo and fixed failing tests
