+++ cmd/mothership/mothership/start/scheduler.go
@@ -17,7 +17,7 @@ func startScheduler(ctx context.Context, o *Options, configFile string) error {
 		return err
 	}
 
-	runtimeBuilder := service.NewRuntimeBuilder(o.Registry.ReconciliationRepository(), logger.NewLogger(o.Verbose))
+	runtimeBuilder := service.NewRuntimeBuilder(o.Registry.ReconciliationRepository(),o.Registry.WorkerRepository(), logger.NewLogger(o.Verbose))
 
 	return runtimeBuilder.
 		RunRemote(
