+++ internal/persistency/registry.go
@@ -6,6 +6,7 @@ import (
 	"github.com/kyma-incubator/reconciler/pkg/kv"
 	"github.com/kyma-incubator/reconciler/pkg/logger"
 	"github.com/kyma-incubator/reconciler/pkg/metrics"
+	"github.com/kyma-incubator/reconciler/pkg/scheduler/occupancy"
 	"github.com/kyma-incubator/reconciler/pkg/scheduler/reconciliation"
 	"go.uber.org/zap"
 )
@@ -16,8 +17,9 @@ type Registry struct {
 	connection      db.Connection
 	inventory       cluster.Inventory
 	kvRepository    *kv.Repository
-	reconRepository reconciliation.Repository
-	initialized     bool
+	reconRepository  reconciliation.Repository
+	workerRepository occupancy.Repository
+	initialized      bool
 }
 
 func NewRegistry(cf db.ConnectionFactory, debug bool) (*Registry, error) {
@@ -46,6 +48,9 @@ func (or *Registry) init() error {
 		return err
 	}
 	if or.reconRepository, err = or.initReconciliationRepository(); err != nil {
+		return err
+	}
+	if or.workerRepository, err = or.initWorkerRepository(); err != nil {
 		return err
 	}
 
@@ -77,6 +82,10 @@ func (or *Registry) ReconciliationRepository() reconciliation.Repository {
 	return or.reconRepository
 }
 
+func (or *Registry) WorkerRepository() occupancy.Repository {
+	return or.workerRepository
+}
+
 func (or *Registry) initRepository() (*kv.Repository, error) {
 	repository, err := kv.NewRepository(or.connection, or.debug)
 	if err != nil {
@@ -100,4 +109,12 @@ func (or *Registry) initReconciliationRepository() (reconciliation.Repository, e
 		or.logger.Errorf("Failed to create reconciliation repository: %s", err)
 	}
 	return reconRepo, err
+}
+
+func (or *Registry) initWorkerRepository() (occupancy.Repository, error) {
+	workerRepo, err := occupancy.NewPersistentWorkerRepository(or.connection, or.debug)
+	if err != nil {
+		or.logger.Errorf("Failed to create worker repository: %s", err)
+	}
+	return workerRepo, err
 }
